Fixed uninstall code so that it checks if files exist rather than constantly throwing and catching exceptions (much safer). Added new feature: InstantExtraction. The download process can be pipelined into the extraction process. Can significantly reduce install time. We are the first modpack to have this feature.

+++ RelicModManager/InstallerComponents/Installer.cs
@@ -54,6 +54,7 @@ namespace RelhaxModpack
         private List<string> originalPatchNames;
         private FileStream fs;
         private string InstalledFilesLogPath = "";
+        private object lockerInstaller = new object();
 
         private static readonly Stopwatch stopWatch = new Stopwatch();
 
@@ -399,8 +400,11 @@ namespace RelhaxModpack
                 {
                     try
                     {
-                        File.SetAttributes(line, FileAttributes.Normal);
-                        Directory.Delete(line);
+                        if(Directory.Exists(line))
+                        {
+                            File.SetAttributes(line, FileAttributes.Normal);
+                            Directory.Delete(line);
+                        }
                     }
                     catch       // catch exception if folder is not empty
                     { }
@@ -409,12 +413,21 @@ namespace RelhaxModpack
                 {
                     try
                     {
-                        File.SetAttributes(line, FileAttributes.Normal);
-                        FileAttributes attr = File.GetAttributes(line);
-                        if (attr.HasFlag(FileAttributes.Directory))
-                            Directory.Delete(line);
-                        else
-                            File.Delete(line);
+                        //always try to solve the problem without throwing exceptions
+                        //https://stackoverflow.com/questions/1395205/better-way-to-check-if-a-path-is-a-file-or-a-directory
+                        if ((File.Exists(line)) || (Directory.Exists(line)))
+                        {
+                            File.SetAttributes(line, FileAttributes.Normal);
+                            FileAttributes attr = File.GetAttributes(line);
+                            if (attr.HasFlag(FileAttributes.Directory))
+                            {
+                                Directory.Delete(line);
+                            }
+                            else
+                            {
+                                File.Delete(line);
+                            }
+                        }
                     }
                     catch (DirectoryNotFoundException)
                     {
@@ -727,7 +740,15 @@ namespace RelhaxModpack
                         Utils.AppendToLog("Extracting Global Dependency " + d.ZipFile);
                         try
                         {
-                            this.Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
+                            if(Settings.InstantExtraction)
+                            {
+                                lock (lockerInstaller)
+                                {
+                                    while (!d.ReadyForInstall)
+                                        System.Threading.Thread.Sleep(50);
+                                }
+                            }
+                            Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
                             args.ParrentProcessed++;
                         }
                         catch (Exception ex)
@@ -752,7 +773,15 @@ namespace RelhaxModpack
                         Utils.AppendToLog("Extracting Dependency " + d.ZipFile);
                         try
                         {
-                            this.Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
+                            if (Settings.InstantExtraction)
+                            {
+                                lock (lockerInstaller)
+                                {
+                                    while (!d.ReadyForInstall)
+                                        System.Threading.Thread.Sleep(50);
+                                }
+                            }
+                            Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
                             args.ParrentProcessed++;
                         }
                         catch (Exception ex)
@@ -777,7 +806,15 @@ namespace RelhaxModpack
                         Utils.AppendToLog("Extracting Logical Dependency " + d.ZipFile);
                         try
                         {
-                            this.Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
+                            if (Settings.InstantExtraction)
+                            {
+                                lock (lockerInstaller)
+                                {
+                                    while (!d.ReadyForInstall)
+                                        System.Threading.Thread.Sleep(50);
+                                }
+                            }
+                            Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
                             args.ParrentProcessed++;
                         }
                         catch (Exception ex)
@@ -804,7 +841,15 @@ namespace RelhaxModpack
                         Utils.AppendToLog("Extracting Mod/Config " + dbo.ZipFile);
                         try
                         {
-                            this.Unzip(Path.Combine(downloadedFilesDir, dbo.ZipFile), TanksLocation);
+                            if (Settings.InstantExtraction)
+                            {
+                                lock (lockerInstaller)
+                                {
+                                    while (!dbo.ReadyForInstall)
+                                        System.Threading.Thread.Sleep(50);
+                                }
+                            }
+                            Unzip(Path.Combine(downloadedFilesDir, dbo.ZipFile), TanksLocation);
                             args.ParrentProcessed++;
                         }
                         catch (Exception ex)
@@ -829,7 +874,15 @@ namespace RelhaxModpack
                         Utils.AppendToLog("Extracting Appended Dependency " + d.ZipFile);
                         try
                         {
-                            this.Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
+                            if (Settings.InstantExtraction)
+                            {
+                                lock (lockerInstaller)
+                                {
+                                    while (!d.ReadyForInstall)
+                                        System.Threading.Thread.Sleep(50);
+                                }
+                            }
+                            Unzip(Path.Combine(downloadedFilesDir, d.ZipFile), TanksLocation);
                             args.ParrentProcessed++;
                         }
                         catch (Exception ex)
